feat: NLW Spacetime - Aula 03

+++ server/src/routes/memories.ts
@@ -1,78 +1,99 @@
-import { FastifyInstance } from "fastify";
-import { z } from "zod";
+import { FastifyInstance } from 'fastify'
+import { z } from 'zod'
 
-import { prisma } from "../lib/prisma";
+import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
-      orderBy: {
-        createdAt: "asc",
+      where: {
+        userId: request.user.sub,
       },
-    });
+      orderBy: {
+        createdAt: 'asc',
+      },
+    })
 
     return memories.map((memory) => {
       return {
         id: memory.id,
         coverUrl: memory.coverUrl,
-        excerpt: memory.content.substring(0, 115).concat("..."),
-      };
-    });
-  });
+        excerpt: memory.content.substring(0, 115).concat('...'),
+      }
+    })
+  })
 
-  app.get("/memories/:id", async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
-    });
+    })
 
-    return memory;
-  });
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
-  app.post("/memories", async (request) => {
+    return memory
+  })
+
+  app.post('/memories', async (request) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    });
+    })
 
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: "83d07b1b-056a-4a7a-9c12-f4a82e9520b0",
+        userId: request.user.sub,
       },
-    });
+    })
 
-    return memory;
-  });
+    return memory
+  })
 
-  app.put("/memories/:id", async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
 
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    });
+    })
 
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -81,22 +102,32 @@ export async function memoriesRoutes(app: FastifyInstance) {
         coverUrl,
         isPublic,
       },
-    });
+    })
 
-    return memory;
-  });
+    return memory
+  })
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
         id,
       },
-    });
-  });
+    })
+  })
 }
